FIX multiple imports of Dancer
Before this patch, it was not possible to import Dancer more than once,
because of the compiled symbols (see the _get_orig_symbol() function for
details).

Also added here some debug infos to help us track what's happening when the app
is starting.

Note that multiple apps dont work yet, the server needs to be aware of them
now, this is coming in the next commit, hopefully ;)