add class Dancer::Core::Dispatcher
There was missing a semantic class in the core: the dispatcher. This class now
provides that feature, making it more clean in the Server role, the psgi_app
now looks dead-simple:

    sub {
        my $env = shift;
        $self->dispatcher->dispatch($env);
    };

Also another benefit is that it's now very easy to test the dispatching without
needing a whole HTTP stack.

And as mst told on IRC, this will open a door for multiple dispatching rule in
the future.