First version of Dancer::Test
This is going to be so much better than with Dancer1:
Dancer::Test is going to do exactly the same flow as a regular
server: it will create an env and pass it to a dispatcher.

the only job of Dancer::Test is to translate the arguments into a valid
environment.

The response returned is a PSGI response.