Dancer::Handler::File - don't use regexp route so we can mix with prefix
Asaaf Gordon reported on the ML that a fresh new app couldn't use
prefix anymore because the Dancer::Handler::File registered a route
with a regexp pattern (and mixing prefixes and Regexp is not
supported).

This produced the following error:

  Cannot combine a prefix (/testme) with a regular expression
  ((?-xism:.*)) at (eval 72) line 74

Now, the Dancer::Handler::File route is '/**' which provides the same
match rules (it's a catchall actually) and is mixable with a prefix.