Dancer::Test compiles its symbols with the caller $app
Just like Dancer.pm, Dancer::Test now compiles its symbols so that
they receive the caller's application as their first argument.

It's safer than using caller each time a Dancer::Test symbol is
called: this time, the caller is static once the module is imported.

TODO: maybe we should refactor the black-magic part (import stuff)
into a role or something, because it's the same code as we have in
Dancer.pm.