First version of send_file that uses Dancer::Handler::File
This supports basic send_file usage (unstreamed).

The idea was to reuse the code Dancer::Handler::File in order to make
sure we don't repeat the knowledge of headers, file reading
protections, ...

Next step is to support streaming while still keeping the knowlegde of
file reading in Dancer::Handler::File.