FIX request's body parsing hangup
A Context object was created for each application, meaning that the body
parsing was done as many times as the number of apps.

The first time, reading the boddy succeeded, but the next times, it failed,
because the input handle was empty, causing ->read to wait for content.

This is fixed by this commit by making sure we create only one Context object
per dispatching, and thus, one Request object is created, no more.

This is also, obviously, a performance gain.