Pulling in documentation from Dancer.
Getting the .pod documentation from Dancer here, ready to be tweaked where
relevant to suit Dancer2, so we're one step closer.

We'll probably want to write up a doc detailing the 1->2 migration and listing
the changes which may be required to get a Dancer1-powered app running under
Dancer2 - `Dancer::UpgradeNotes` or similar, perhaps.