Support for "postponed hooks" (fix the issue faced in t/hooks.t)
The idea is to register hooks in a "postponed" hash if none of the known
hookbale instances claim the hook.

That postponed stash is shared with the app and the runner object. When
something that is hookable is created, we go over that hash in order to find
some postponed hook for it.

This is done thanks to a nice naming convention that we have now:

Core hookable objects must regsiter their hooks under the "core.$name"
namespace (eg: "core.error.init" or "core.app.before_request").

Engine hooks must go under "engine.$name" (eg: engine.template.before_render)

Plugin hooks must go under "plugin.$name".

Note: the whole test suite pass, and the t/hooks.t script now register template
and serializer hooks before setting them up, and that works. BUT. Some work is
probably still needed for plugins. We'll see.