allow before hook to populate the response's content
This patch is to accomodate plugins that needs to short-circuit
the route's main code (in my case, that's D::P::Cache::CHI).

With this patch, if the 'before' hook halts the response,
the response content is taken from whatever the hook put in
$context->response->content, not from $route->execute($context).
The grooming of the content (serializing, content-type filling,
content encoding), however, is still done.