Support for register_hook in Dancer::Plugin
This will makes a lot easier plugins migration for plugins that
need to register their own hooks.

The register_hook does all the magic of registering the hook with
its long name "plugin.$plugin_name.$hook_name" and its "alias" name.

As a plugin consumes the Hookable role, this is done only by
implementing the functions "supported_hooks" and "hook_aliases".

A new test case is provided to demonstrate that.