Support for config sharing between applications.
The main idea behind this change is to allow an app to pass one or
more config entries to another app, when it 'use's it.

	package MyApp;
	use Dancer 2.0;

	set session => 'YAML';
	use Blog with => { session => engine('session') };

In this example, the session engine is passed to the "Blog" app.
That way, anything done in the session will be shared between both
apps.

Anything that is defined in the config entry can be passed that way.

If we want to pass the whole config object, it can be done like so:

	use SomeApp with => { %{config()} };