Dancer::Session::YAML doesn't store its whole $self in files
Before this patch, the content dumped into the YAML file was the whole
content of the session object.

This was overkill, not needed, and looked to create a bug when sharing
the same engine.

Now, the session dumped into the file is the "data" of the instance, not
the instance itself.