Fix: Plugins can be used by more than one app.
Dancer::Plugin was broken: its "register_plugin" method was doing two
kinds of things:
  - registering the plugin (init stuff for the plugin)
  - exporting the keywords plugin to the caller

When there is only one app that uses the plugin, it's fine, because
the register plugin will work as epxected, but when another app uses
the plugin, it fails, because register_plugin won't be called anymore.

With this patch, register_plugin only does the registration stuff,
and the export stuff is done at import time, whenever an app uses the
plugin.

A new test has been added to demonstrate the issue
(t/plugin_exports.t).