FIX UTF-8 response: Content-Length is now properly set
When rendering UTF-8 responses, the Content-Length header was not properly set,
it was half the value of the real strings.

Now, when the response is encoded to UTF-8, the header is reset in order to be
real.

New test script added (taken from Dancer 1).

The charset suffix is also appended to the Content-Type now.