Complete rewrite of the session handling.
Session objects and session engines are now properly decoupled.

The Dancer::Core::Role::SessionFactory role is responsible for backend
implementations whereas the Dancer::Core::Session class implements a pure
session object.

We now have backend-agnostic sessions on one side, and session backends on the
other one.

The session cookie is set once, and only once, in a before filter (which is
automatically added to the application at BUILD time, if a session engine is
set.)

In the same manner, the session "flush" is done once and only once, in an
after filter.

The design is elegant and should be much better performance-wise.

To do things properly, a "session" attribute has been added to the context
object, which is the right place for it. The session attribute of the Context
object is lazy, and its builder is responsible for fetching an existing session
or creating a new one, depending on the presence of a valid dancer.session
cookie in the request.

All these changes make the code much better to read and to maintain.

Also, the bug with multiple sessions created for one request is now solved.