DEPRECATION on direct calls to keywords in plugins
When Dancer2 was compliant with Dancer 1 plugins, it provided a backward-compat
syntax for plugins: the ability to use Dancer's main keyword directly in the
code, just like end-user applications.

In Dancer2 it causes more harm than good and we'll now enforce plugins in Dancer
2 to use the $dsl object each keyword receives as its first argument.

To ease the transition for existing plugins, this commit only issue a
DEPRECATION notice.

The test suite and the Dancer2::Plugin::Ajax were fixed here also.