HEAD request drops the body of the response.
For a HEAD request, the body of the response should be empty ('').  By
implementing an 'around' modifier to the 'dispath' method, we can drop
the body if the request is HEAD.

The reason to use an 'around' modifier in this case, is that the
Content-Length head should contains the correct lenght of the response.

Closes #286.