Fix for the default tests of a new generatd app.
The tests of an app generated by the `dancer2` scripts are failing
because the tests can't load the application.

By specifiying the name of the application to load in the test we can
make them pass.

Also, in the Dancer2::Test module, before calling the 'can' method on
the caller, we check if the caller is defined, that way, instead of
failing the tests because we can't call a method on an undefined object,
we fail with a useful message for the user.

Closes #293