Pass the request object to the Route's `match' method.
Instead of passing only the path and the HTTP method to the `match'
method, we pass the entire request object.  This should be usefull for
matching options if a route is defined with some.

Also, the HTTP method was never used in this method, so this should have
no impact.