Load the views and public directory from the correct place.
Dancer was trying to load the `views' and `public' directories from the
configuration directory. Until now this was correct because the
application's location and configuration's location were always the
same.  Now that it's possible to specify a different location for each
of them, we should try to load `public' and `views' from the
application's location (that was the default behavior in Dancer 1).