Methods to return path of views and layout.
Rename the `view' method to `view_pathname', to make it more explicit
about the meaning of that method.  Also add a `layout_pathname' method,
that returns the full path to the layout.

Add documentation for the role so other developpers can implement
different plugins to support more templating engines.

This is a reimplementation of the work done by Yannick on #32.