Implement the code to support the 'startup_info' directive.
By default, if the environment variable DANCER_STARTUP_INFO is not set,
and the setting 'startup_info' is also not set, silence the banner from
HTTP::Server::Simple.  In the case one of the setting is true, display
Dancer's banner with some additional informations.

This is based on the work done by Maurice Mengel.

Closes #210