Create methods to build the various engines.
Since we support only four engines, we can have dedicated methods that
we call and store their values in the attribute 'engines'.

We *should* never store the object in the config, and therefore, everytime
someone need an engine, it should be done via '->engine($name)', and never
via '->settings->{$name}'.

This is still broken.