Move the triggers in the config to proper attributes.
There's two kinds of triggeres: on engines, and on config change.

We add a new attribute 'supported_engines' that contains a list of all the
supported engines by Dancer2.  In addition to that, two private attributes are
also added, one that contains a list of triggers associated to the engines,
and the second one is for the config change.