Drop support for log_path in Dancer2::Logger::File.
The log_path option was available as a backward compatibility feature
in Dancer1, but there's no need to cary this one over Dancer2.

This change remove the mention of log_file and log_path from the documentation,
and makes the default location for the log file (if none specified) to be the
application's directory.