remove trigger in caller to set location:
In order to remove the trigger, we need the builder to actually
be a real builder: run for the creation of the attribute and returning
the location string back.

The problem is that the attribute comes from the Config role, and
it has a default there. Instead, we remove the default and replace
it with a builder. Then we can override the builder in the runner
so it has its own way of building the location.

The "required" is removed because it's pointless to have a "required"
when there's a default or builder. It will always have a value.