This is just silly:
The "BUILDARGS" method can be used to set/delete/change values that
will be used to generate the class. Running "around" on it is less
than pointless since that's what "BUILDARGS" is used for.

More interestingly, this tries to accomplish a default value for the
attribute which can be set via the "default" option. More interestingly,
there is *already* a default attribute for it set.