Don't futz with $^W. Import warnings and remove import_warnings conf setting.
Use warnings->import when importing Dancer2 to ensure any code using it has
the warnings pragma enabled (just like strict and utf8).

L<perllexwarn> states ".. using $^W to either disable or enable blocks of
code is fundamentally flawed.". Also see Matt Trout's comments in #427.

Remove the import_config setting and its trigger that mucks with $^W and purge
it from the POD. If anyone really wants to set $^W, they can add it to their
app or psgi file and be in control of the results.