Force YAML::XS
The *::Any modules are a lie. Whenever one of the backends has a bug it is propagated to the Any module, and we need to deal with it. As pointed in GH#433, without YAML::XS there are encoding problems, and deprecation warnings. My feeling is that is is better to require the real thing.