Tested session hooks for Simple engine as per #383.
This test suite still only works with Simple and not with more than one
backend. Possibly because I'm not setting it all up correctly. This
changeset also added LWP::Protocol::PSGI as a dependency into dist.ini.
I needed this to setup the dancer app and make requests to it within the
same process.