Make default_config an attribute allowing overrides during construction.
The default_config in the runner forms the basis of what should be immutable
configuration for other objects. Allow these to be overridden during
construction with the aim at making as many of these attributes read-only as
possible later.

If a default_config hashref is supplied during construction, its merged with
the existing defaults from the lazy builder for default_config.

Only watch was there were two being build based on the Runners location. These
are now populated in the BUILD method.