GH#490: Document Dancer2 should be fatpackable:
In Dancer2 we can optionally use XS modules for speed boosts, but we
should remain aware that it is still meant to be fatpackable. I could
not think of a more proper place for this statement than in the main
POD, but feel free to suggest better locations and move it.