Move POD to end, fix UTF8 (RT #91428):
I hate documentation in the middle of code. I can't concentrate
on the code itself. Couple that with the way POD::Weaver works,
merging documentation patches is awkward and non-trivial.

I moved all the POD to the end, after the __END__ marker (ha!), and
merged a fix by Gregor Herrmann to declare UTF-8. This was necessary
because we now include UTF-8 characters under the list of names
(core team, contributors).