Add a default_content_type attribute to the Response object.
Part I in refactoring default_content_type attribute from the Dispatcher
to the Response object. This feels a better fit; the only downside is there
is no back references to the context or app from the response.

Update encode_content() and to_psgi() to add this default_content_type
to the response if no content_type was previously specified.