Use Plack middleware for HEAD request content removal.
The logic in the around => dispatch modifier required a request to be passed in
before it would remove content from the response of a HEAD request. All the
current Servers only supply the psgi env and no content removal was occuring.
(Plack::Middleware::Head is generic middleware that handles exactly this case)

Note that previously Dancer2::Test was supplying both the env and a request
causing the test to pass. Another subtle win for moving to Plack::Test.