Untangle Runner->Server, use HTTP::Server::PSGI:
Runner would create a server, of type PSGI or Standalone,
which would then either return a PSGI app coderef or run a
local server, respectively.

Beyond how silly it is to create a "PSGI server" that only returns
a PSGI coderef, and a "Standalone server" which is actually a PSGI
server, this is superfluous.

There was much confusion on which attributes (postponed_hooks, apps)
are kept in which class, and this became worse when deciding how to
propagate them to the app and back.

These are now all done in the runner, and the server has been
reduced to an attribute in the Runner.

The server has also been replaced with HTTP::Server::PSGI to remove
HTTP::Server::Simple::PSGI from our dependency chain and our minds.