Make the list of apps to dispatch to a param to Runner->psgi_app
If you build a larger app from several smaller independent Dancer2 apps
that have similar routes (such as the existing Cookbook OurForum/OurWiki
example), you need a separate list of apps to dispatch over for the psgi
app for each, otherwise the first matching route will always win.

Allow a list of apps to be specified to dispatch over for a specific
psgi coderef. If non is specified, the default remains to dispatch over
all apps.