Remove default engine configuration:
Default engine configurations are configurations that are
automatically sent to every engine. This makes no sense, because
each engine will have its own configuration.

An example would be a template engine that has a template object
inside it. The template object gets sent the default configuration.

The default configuration contains two keys - environment and
location - which either means nothing to the template object, or
it means something probably different.

Instead, these should be sent as attributes to the engines. They
can be configured in the Engine role and automatically fed. Thing
is, I couldn't find anything that actually uses them and needs
them, so I removed it completely.