Move environment/location from Engine role to Logger::File, give defaults:
Dancer2::Logger::File actually needs the attributes environment/location
which we removed in this branch. It is set in the Engine role.

This commit does two things:
1. Put those attributes in the File logger module
2. Give them the proper values using the App via the Context

The 2nd item is the most horrible way of doing it. I'm not sure how else
to do it, but it will need to be changed as part of the Context decoupling
work which is something we definitely want to do.