Make apphandler a global variable:
Recently when running

    set apphandler => 'PSGI'

it did not pick up because setting the apphandler affected only
the Dancer App, while the apphandler was (correctly) decided and
used by the Dancer Runner.

By making the "apphandler" option a global trigger that updates the
Dancer Runner object, this will now work.

Added a test that uses LWP::Protocol::PSGI, which exposed this
problem in GH #447.