Make behind_proxy a global var
Conditionally apply BehindProxy middleware if behind_proxy is set.

Making the "behind_proxy" option a global trigger that updates the
Dancer 2::Runner object when
  set behind_proxy => 1;
is run so the reverse proxy middleware can be conditionally wrapped
around applications psgi coderef.

We now 'use' the middleware directly rather than letting Placl::Builder
load it for us. May allow earlier version of Placl to be used as a dependency.