Remove the _build_request method from Core::Context:
In an effort to find more code that bypasses Core::Dispatcher,
we removed the code that creates the request internally in
Core::Context.

This showed:
* A check in response_not_found method that generates a 404, which
  needs to set the new request it generates in Core::Context.
* A test (context.t) which abuses Core::Context in order to check
  its variables (creating a context and request without an app).
  This has been fixed with an ugly workaround to get moving. This
  test will eventually be completely removed when Core::Context is.
* A test (error.t) which is also using Core::Context directly, which
  it shouldn't.

In the future, any test that uses Core::Context will need to be
rewritten, since it's clearly abusing the internals instead of going
through the normal path to get a PSGI coderef, or checking the
objects cleanly.