Moved the rest of session related attributes and methods to Core::App
'session' attribute was removed from Core::Context

Moved from Core::Context to Core::App:
* destroyed_session (attribute)
* _build_session (replaced the temp. writer)
* has_session
* destroy_session

Added the clearer 'clear_destroyed_session' to 'destroyed_session'
attribute and now it runs in app_cleanup.
(this is to make sure we don't keep a destroyed session for the
next request -- that was a fun debugging session)

Dispatcher now takes curr_session instead of curr_context as optional
parameter (in forward).

Some docs references to Context where changed to App.

This is a major step forward in the effort to remove Core::Context
(almost there)