GH #587: Port Serializer::Mutable
Serializer::Mutable is available on D1 but not D2.
No more!

The following helper DSL keywords would shift and take app, even
though it wasn't necessary, and it broke our ability to work with
them (whether we were "supposed to" work with them or not). Because
it isn't actually using the app, we replaced them with "shift":

* from_json
* to_json
* from_yaml
* to_yaml
* from_dumper
* to_dumper