BugFix #336: 'set log => ...'  doesn't set the log level
Reported originally by @andrewsolomon. Dancer2::Tutorial states
that set log => warning, should set the log_level accordingly.
However, this shortcut was not working and it was necessary
to directly access the log engine like this:
    setting('logger')->{'log_level'} = 'warning'

This sets it staight by creating a new App trigger 'log' which
will set log_level as expected.