Collect caller/location attrs into single role.
Currently Dancer2::Runner defineds the caller app and the builder for
location, while the location attribute is part of the ConfigReader role.
Collect them all together into a role that provides the location
attribute.

One downside is to the current split is that the caller is set on first
import Dancer2, which gets in the way of supporting multiple apps with
separate configs. This is also some of the most unliked code in the
codebase; its now all together for future revision...