Remove references to caller and location.
The Runner is a singleton; it should not have the configuration of
the first app that imports Dancer2. Start by removing its caller and
location attributes (i.e. do not include the ConfigReader role).

However the Runner does require to manage some global configuration
for all apps; such as the environment and properties of the default
server (port, host, etc.). Add a 'config' attribute to keep track
of this (and somewhere for global triggers to update).