Use 'execute' usable
In case an external application wants to bootstrap a Dancer2 module it
has two options, use `dancer2` command, or try to use the CLI gen
command directly.

To use the gen command directly it would be great if the execute method
could be called without a self (thus, not needing to prepare an
App::Cmd::Command object).

Given `_version_check` and `version` really do not use the object, we
can remove it.

An alternative, if you prefer, would be to take most of the code from
the `execute` method and create a standalone method one can use.

Ideas?