Return default for caller, don't confuse scopes:
The default for the "caller" attribute was necessary because it
wasn't provided in various test scripts. However, it cannot be
written as a sub() in and of itself, because it then enters another
scope. By entering another scope, it overrides the caller with the
value "(eval <number>)", so instead, we use Sub::Quote in order to
make that inlined, thus retaining the original scope.

A problem that emerged is the shitty t/app.t test which checked
for $! instead of $@ after an eval. Because of this problem of scopes
a previous piece of code in that script failed and raised $!, which
was checked (instead of $@) to conclude a check there failed as it
should.

As soon as we fixed the scope problem with Sub::Quote, it suddenly
didn't raise a system error and $! became undef. That's when the
additional check (which should have checked $@) failed. We fixed
that too.

More context on this can be found at:

https://github.com/PerlDancer/Dancer2/pull/611