Explictly state which 'caller' we require.
There was an ambiguity between which 'caller' that was implied within
Role::HasLocation. Was it the accessor for the 'caller' attribute, or
the core function 'caller'. The following warning was being generated:
"Ambiguous call resolved as CORE::caller(), qualify as such or use &"

Explictly use CORE::caller to remove the ambiguity in the default for
the caller attribute.