Add a cookie object to the new request on forward.
If a session was inflated form an existing Cookie, the logic in the
dispatcher will propogate that into apps' context on forward. However if
the session object was created during the current request, that object
has no information as to which app(s) it belonged to ( for which the
session engines' cookie_name is a proxy for)

Add a cookie to the new request if one didn't already exist for the
apps' cookie name so the dispatcher can propogate the session if
required.