Add test for another possible case:
As explained in GH #638, because we're only storing the last session,
we might clobber a session with its data that was forwarded twice.

This happens if the forwarding was done to apps with different
cookie_name.

Imagine you have three applications: App1, App2, and App3.

App1 and App3 use the same session cookie_name variable. App2 has
a different one.

Now imagine a single request is being dispatched to App1, then
forwarded to App2, and then forwarded to App3. Each one wants to
access the session for something.

Since App1 and App3 use the same session cookie_name, they expect each
of them to be able to reach the data they put it and the data the other
one put it, but *not* the data App2 (which has a different session
cookie_name variable) has, because it's stored under a different ID
altogether.

Here is the order of events:
* Reach App1
* Using 'session()' will create a new session and store data in it
* Before forwarding to App2, save the current session
* Before dispatching to App2, check if it's the same cookie_name
* It isn't, then dispatch to it without the session from App1
* Using 'session()' in App2 will create a new session
* Before forwarding to App3, save the current session

  (* We just overwrote the session of App1 with the one from App2 *)

* Before dipsatching to App3, check if it's the same cookie_name
* It isn't, then dipsatch to it without the session from App2
* App3 doesn't know about the data from App1's session,
  even though it's using the same session cookie_name
  and should have the same session data

This test checks for it.