Test for template engine hooks overloading in multiple apps:
An issue Mickey and I noticed was that the wrong before_template
hook was called.

We had:
- App1
- App2

We loaded both apps, and then making a request to App2 specifically,
which then triggered the before_template hook from App1 instead.

After writing this test script, it would appear (and it's now
expressed in the test as well) that App2 gets both the before_template
which is defined in it, but also the one from App1.

Further investigation required to find why this happens, although I have
my suspects.