Postponed hooks by app name:
All of the postponed hooks are available through the runner and later
composed into each App. The problem is that the App receives all the
postponed hooks that were registered in the Runner, even if they don't
belong to it.

That means that every App receives every postponed hook (like
before_template) that every previous App created. App1's postponed
hooks are sent to App2 (along with any postponed hooks App2 has)
and all of them will be sent to App3 (along with any postponed hooks
App3 might have).

t/multiapp_template_hooks.t shows this.

The problem is solved by simply using the application name in order
to store the postponed hooks in the Runner to make sure we only add
the postponed hooks that are relevant to the App.